pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-score"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "scikit-learn",
    "scipy",
    "numpy",
    "pandas",
    "requests",
    "pyyaml",
    "mlflow>=2.0",
]

[project.scripts]
risk-score = "risk_score.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
